leaf/leaf/requirements.txt
cycler==0.9.0
Django==1.8.8
django-appconf==1.0.1
django-geoposition==0.2.2
matplotlib==1.5.1
numpy==1.10.4
Pillow==3.1.0
pyparsing==2.0.7
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
wheel==0.24.0
gunicorn
